+++ src/components/Select/index.ts
@@ -1,3 +1,3 @@
-// Select from copy-ui @ 2025-11-24.
+// Select from copy-ui @ 2025-11-28.
 
 export { Select } from "./select";