+++ src/components/Select/CHANGELOG.md
@@ -18,3 +18,4 @@
   `min-width` constraint on the trigger. And remove the dependency on
   `ThemeProvider`.
 - 2025-11-24: Support the empty state.
+- 2025-11-28: Support the `xs` and `xl` width props.